requirements.txt
# Web Framework
streamlit>=1.28.0

# NLP & Embeddings
sentence-transformers>=2.2.0

# PDF Processing
pymupdf>=1.23.0

# Numerical Computing
numpy>=1.24.0

# LLM API
openai>=1.0.0

# Environment Variables
python-dotenv>=1.0.0